[FEAT] 유형 테스트 로직 변경 및 나의 나나에서도 테스트 다시하기 구현
