+++ NanaLand/NanaLand/Common/Util/AuthManager.swift
@@ -187,7 +187,7 @@ final class AuthManager: NSObject {
 				providerId: request.providerId
 			)
 			
-			registerVM.state.isRegisterNeeded = true
+			AppState.shared.isRegisterNeeded = true
 		}
 	}
