+++ NanaLand/NanaLand/Common/ViewType/RegisterViewType.swift
@@ -1,0 +1,13 @@
+//
+//  RegisterViewType.swift
+//  NanaLand
+//
+//  Created by 정현우 on 5/24/24.
+//
+
+import Foundation
+
+enum RegisterViewType {
+	case termsAgreement  // 약관 동의
+	case nicknameAndProfile  // 닉네임과 프로필사진
+}
